blah

+++ lib/lib/src/index.ts
@@ -1,5 +1,5 @@
 import { double } from "./module"
 
-const quadruple = (value: number) => double(double(double(value)))
+const quadruple = (value: number) => double(double(value))
 
 export { quadruple }
